refactor: clean router.py and service.py in /employees/

+++ employees/service.py
@@ -2,14 +2,14 @@ from sqlalchemy.ext.asyncio import AsyncSession
 
 import employees.repository as repository
 import departments.repository as department_repository
-from employees.schema import AddressCreate, EmployeeCreate, EmployeePatch
-from exceptions.handler import BadRequestException, NotFoundException
-from models import employee
+from employees.schema import AddressCreate, AddressPatch, EmployeeCreate, EmployeePatch
+from exceptions.handler import NotFoundException
+from models import address
 from models.address import Address
 from models.employee import Employee
 
 
-async def create_employee( body: EmployeeCreate, db: AsyncSession):
+async def create_employee( body: EmployeeCreate, db: AsyncSession) -> Employee:
   employee: Employee = Employee()
 
   employee.name = body.name.strip()
@@ -35,8 +35,7 @@ async def get_all_employees(db: AsyncSession) -> list[Employee]:
 
 async def get_employee( employee_id: int, db: AsyncSession) -> Employee:
   employee = await repository.get_employee(employee_id, db)
-
-  if employee is None:
+  if employee is None or employee.deleted_at is not None:
     raise NotFoundException(detail="Employee not found in DB")
 
   return employee
@@ -54,13 +53,13 @@ async def delete_employee(employee_id: int, db: AsyncSession)  -> Employee:
 async def patch_employee(employee_id: int, body: EmployeePatch, db: AsyncSession, ) -> Employee:
   original_employee: Employee = await repository.get_employee(employee_id, db)
 
-  if original_employee is None:
+  if original_employee is None or original_employee.deleted_at is not None:
     raise NotFoundException("Requested employee is not present in the DB")
 
   if body.name is not None:
-    original_employee.name = body.name
+    original_employee.name = body.name.strip()
   if body.email is not None:
-    original_employee.email = body.email
+    original_employee.email = body.email.strip()
   if body.age is not None:
     original_employee.age = body.age
 
@@ -68,7 +67,7 @@ async def patch_employee(employee_id: int, body: EmployeePatch, db: AsyncSession
   return patched_employee
   
 
-async def create_address(employee_id: int, body: AddressCreate, db: AsyncSession):
+async def create_address(employee_id: int, body: AddressCreate, db: AsyncSession) -> Address:
   employee = await repository.get_employee(employee_id, db)
 
   if employee is None or employee.deleted_at is not None:
@@ -82,57 +81,67 @@ async def create_address(employee_id: int, body: AddressCreate, db: AsyncSession
 
   return await repository.create_address(employee, address, db)
 
-async def get_address(address_id: int, db: AsyncSession):
+async def get_address(address_id: int, db: AsyncSession) -> Address:
   address: Address = await repository.get_address(address_id, db)
   
-  if address is None:
+  if address is None or address.deleted_at is not None:
     raise NotFoundException(detail="The given address was not found in the DB")
   
   return address
 
-async def delete_address(address_id: int, db: AsyncSession):
+async def delete_address(address_id: int, db: AsyncSession) -> Address:
   address: Address = await repository.get_address(address_id, db)
 
-  if address is None:
+  if address is None or address.deleted_at is not None:
     raise NotFoundException(detail="The given address was not found in the DB")
   
   deleted_address: Address = await repository.delete_address(address, db)
 
   return deleted_address
   
-async def patch_address(address_id: int, body: EmployeePatch, db: AsyncSession):
+async def patch_address(address_id: int, body: AddressPatch, db: AsyncSession) -> Address:
   original_address = await repository.get_address(address_id, db)
 
-  if original_address is None:
+  if original_address is None or original_address.deleted_at is not None:
     raise NotFoundException(detail="The given address was not found in the DB")
   
   if body.street is not None:
-    original_address.street = body.street
+    original_address.street = body.street.strip()
   if body.city is not None:
-    original_address.city = body.city
+    original_address.city = body.city.strip()
   if body.country is not None:
-    original_address.country = body.country
+    original_address.country = body.country.strip()
   if body.postal_code is not None:
-    original_address.postal_code = body.postal_code
+    original_address.postal_code = body.postal_code.strip()
 
   patched_address = await repository.patch_address(original_address, db)
   return patched_address
 
-async def get_all_addresses(employee_id: int, db: AsyncSession):
-  employees = await repository.get_all_addresses(employee_id, db)
-  return employees
+async def get_all_addresses(employee_id: int, db: AsyncSession) -> list[Address]:
+  addresses = await repository.get_all_addresses(employee_id, db)
+  return addresses
 
 
-async def add_department_to_employee( employee_id: int, department_id: int, db: AsyncSession):
+async def add_department_to_employee( employee_id: int, department_id: int, db: AsyncSession) -> Employee:
   employee = await repository.get_employee_with_department(employee_id, db)
   department = await department_repository.get_department(department_id, db)
+
+  if employee is None or employee.deleted_at is not None:
+    raise NotFoundException("Employee not found")
+  if department is None or department.deleted_at is not None:
+    raise NotFoundException("Department not found")
 
   employee.departments.append(department)
   return await repository.add_department_to_employee(employee, db)
 
-async def delete_department_from_employee(employee_id: id, department_id: int, db: AsyncSession):
+async def delete_department_from_employee(employee_id: int, department_id: int, db: AsyncSession) -> Employee:
   employee = await repository.get_employee_with_department(employee_id, db)
   department = await department_repository.get_department(department_id, db)
+
+  if employee is None or employee.deleted_at is not None:
+    raise NotFoundException("Employee not found")
+  if department is None or department.deleted_at is not None:
+    raise NotFoundException("Department not found")
   
   employee.departments.remove(department)
   return await repository.delete_department_from_employee(employee, db)
